[GROW-262] update file architecture
