+++ Planz/AppView.swift
@@ -25,22 +25,3 @@ struct AppView: View {
         }
     }
 }
-struct PromiseContentView: View {
-    var store: Store<MakePromiseState, MakePromiseAction>
-    let selectThemeStore: Store<SelectThemeState, SelectThemeAction> = Store(initialState: SelectThemeState(), reducer: makePromiseSelectThemeReducer, environment: SelectThemeEnvironment())
-    
-    public var body: some View {
-        WithViewStore(self.store) { viewStore in
-            HStack {
-                switch viewStore.currentStep {
-                case .selectTheme:
-                    SelectThemeView(store: selectThemeStore)
-                case .fillNAndPlace:
-                    NameAndPlaceInputView()
-                case .error:
-                    Spacer()
-                }
-            }
-        }
-    }
-}
